Package structure refactoring/cleanup in geomesa-trino-plugin
