+++ geomesa-trino/geomesa-trino-plugin/src/test/java/org/locationtech/geomesa/trino/spatial/iceberg/connector/SpatialConnectorMetadataTest.java
@@ -6,7 +6,7 @@
  * https://www.apache.org/licenses/LICENSE-2.0
  ***********************************************************************/
 
-package org.locationtech.geomesa.trino.spatial.connector;
+package org.locationtech.geomesa.trino.spatial.iceberg.connector;
 
 import io.trino.spi.connector.*;
 import io.trino.spi.expression.*;
