+++ geomesa-trino/geomesa-trino-plugin/src/test/java/org/locationtech/geomesa/trino/spatial/iceberg/transforms/Z2SpatialTest.java
@@ -8,7 +8,6 @@
 
 package org.locationtech.geomesa.trino.spatial.iceberg.transforms;
 
-import org.locationtech.geomesa.iceberg.spatial.transforms.SfcBridge;
 import org.junit.jupiter.api.Test;
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.Envelope;
